pyproject.toml
[project]
name = "kliq-growth-engine"
version = "0.1.0"
description = "KLIQ Growth Engine - Automated coach discovery, webstore generation, and outreach"
requires-python = ">=3.11"
dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Task Queue
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "alembic>=1.13.1",

    # Scraping
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "youtube-transcript-api>=0.6.2",
    "google-api-python-client>=2.114.0",
    "google-auth>=2.27.0",
    "colorthief>=0.2.1",
    "httpx>=0.26.0",

    # AI
    "anthropic>=0.43.0",
    "jinja2>=3.1.3",

    # Email (Brevo)
    "sib-api-v3-sdk>=7.6.0",

    # Cloud
    "boto3>=1.34.0",
    "google-cloud-bigquery>=3.14.0",

    # Auth
    "bcrypt>=4.1.2",

    # Utils
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
